ETL/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.1
orjson==3.8
pydantic==1.9.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
requests==2.28.1
uvicorn==0.12.2
uvloop==0.17.0
gunicorn==20.1.0
httptools==0.5.0